Use Next.JS to build the static site
Replaces our express handling of static-site assets (formerly built with
Gatsby) to use Next.JS. As we are using Next.JS without any
server-generated content the implementation is relatively
straight-forward. Note that when NODE_ENV is _not_ production then
Next.JS will automatically watch for changes to the front-end code and
implement hot-reloading. This makes for a significantly better dev
experience than having to run a separate Gatsby server in development
mode and use a proxy to join our server + Gatsby.
