+++ src/routing/auspice.js
@@ -10,11 +10,11 @@ export function setup(app) {
    * so we must use that prefix here too.
    */
   app.route("/dist/*")
-    .all(endpoints.static.auspiceAssets, (req, res, next) => next(new NotFound()));
+    .all(endpoints.auspice.auspiceAssets, (req, res, next) => next(new NotFound()));
 
   /* Auspice has a special /edit/narratives route -
    * It is not backed by a dataset, and only exists for GET requests
    */
   app.routeAsync("/edit/narratives")
-    .getAsync(endpoints.static.sendAuspiceEntrypoint);
+    .getAsync(endpoints.auspice.sendAuspiceEntrypoint);
 }
