+++ src/routing/community.js
@@ -21,16 +21,18 @@ export function setup(app) {
     setSource(req => new CommunitySource(req.params.user, req.params.repo)));
 
   /* Unlike other dataset and narrative routes, /community/:user/:repo and
-   * /community/narratives/:user/:repo can go to either Auspice or Gatsby,
+   * /community/narratives/:user/:repo can go to either Auspice or Next.JS,
    * depending on the existence of a top-level dataset or narrative file.  The
    * two cases depend on the existence of a dataset file named auspice/:repo.json
    * or narrative file in narratives/:repo.md in the GitHub :user/:repo.
    *
    * If it exists, Auspice's entrypoint is sent (for HTML-accepting requests).
    *
-   * If it doesn't exist, a 404 is raised and (for HTML-accepting requests)
-   * Gatsby's 404 page is sent which then does client-side routing to show
-   * the Gatsby page static-site/src/sections/community-repo-page.jsx.
+   * If it doesn't exist, a NotFound (404) error is raised and (for HTML-accepting
+   * requests) we hand control to the Next.JS app. This is not aware of the
+   * NotFound error and will display the matching page for community. Within that
+   * page (client-side code) we inspect the URL path and display an appropriate
+   * error banner.
    */
   app.routeAsync(["/community/narratives/:user/:repo", "/community/narratives/:user/:repo/*"])
     .all(setNarrative(req => req.params[0]))
